api/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
Flask-SQLAlchemy==3.1.1
kafka-python==2.0.2
marshmallow==3.20.1
jsonschema==4.19.0
requests==2.31.0
python-dotenv==1.0.0
